foundation

+++ src/Foundation/NSValue.cs
@@ -91,7 +91,7 @@ namespace Foundation {
 				// avoid potential buffer overflow since we use the older `getValue:` API to cover all platforms
 				// and we can cheat here with the actual string comparison (since we are the one doing it)
 				if (ObjCType == "{CGAffineTransform=dddddd}")
-					StoreValueAtAddress ((IntPtr) (void*) &result, sizeof (CGAffineTransform));
+					StoreValueAtAddress ((IntPtr) (void*) &result, (nuint) sizeof (CGAffineTransform));
 				return result;
 			}
 		}
